+++ src/sentry/android/android_sdk.cpp
@@ -2,11 +2,13 @@
 
 #include "android_breadcrumb.h"
 #include "android_event.h"
+#include "android_log.h"
 #include "android_string_names.h"
 #include "android_util.h"
 #include "sentry/common_defs.h"
 #include "sentry/logging/print.h"
 #include "sentry/processing/process_event.h"
+#include "sentry/processing/process_log.h"
 #include "sentry/sentry_attachment.h"
 
 #include <godot_cpp/classes/engine.hpp>
@@ -15,7 +17,18 @@
 
 using namespace godot;
 
+namespace {
+
+inline Variant _as_attribute(const Variant &p_value) {
+	Variant::Type type = p_value.get_type();
+	return (type < Variant::BOOL || type > Variant::STRING) ? (Variant)p_value.stringify() : p_value;
+}
+
+} // unnamed namespace
+
 namespace sentry::android {
+
+// *** SentryAndroidBeforeSendHandler
 
 void SentryAndroidBeforeSendHandler::_initialize(Object *p_android_plugin) {
 	android_plugin = p_android_plugin;
@@ -38,6 +51,30 @@ void SentryAndroidBeforeSendHandler::_before_send(int32_t p_event_handle) {
 void SentryAndroidBeforeSendHandler::_bind_methods() {
 	ClassDB::bind_method(D_METHOD("before_send"), &SentryAndroidBeforeSendHandler::_before_send);
 }
+
+// *** SentryAndroidBeforeSendLogHandler
+
+void SentryAndroidBeforeSendLogHandler::_initialize(Object *p_android_plugin) {
+	android_plugin = p_android_plugin;
+}
+
+void SentryAndroidBeforeSendLogHandler::_before_send_log(int32_t p_handle) {
+	Ref<AndroidLog> log_obj = memnew(AndroidLog(android_plugin, p_handle));
+	log_obj->set_as_borrowed();
+
+	Ref<AndroidLog> processed = sentry::process_log(log_obj);
+
+	if (processed.is_null()) {
+		// Discard log.
+		android_plugin->call(ANDROID_SN(releaseLog), p_handle);
+	}
+}
+
+void SentryAndroidBeforeSendLogHandler::_bind_methods() {
+	ClassDB::bind_method(D_METHOD("before_send_log"), &SentryAndroidBeforeSendLogHandler::_before_send_log);
+}
+
+// *** AndroidSDK
 
 void AndroidSDK::set_context(const String &p_key, const Dictionary &p_value) {
 	ERR_FAIL_NULL(android_plugin);
@@ -90,6 +127,28 @@ void AndroidSDK::add_breadcrumb(const Ref<SentryBreadcrumb> &p_breadcrumb) {
 	Ref<AndroidBreadcrumb> crumb = p_breadcrumb;
 	ERR_FAIL_COND(crumb.is_null());
 	android_plugin->call(ANDROID_SN(addBreadcrumb), crumb->get_handle());
+}
+
+void AndroidSDK::log(LogLevel p_level, const String &p_body, const Dictionary &p_attributes) {
+	ERR_FAIL_NULL(android_plugin);
+
+	if (p_body.is_empty()) {
+		return;
+	}
+
+	String body = p_body;
+
+	Dictionary attributes;
+
+	if (!p_attributes.is_empty()) {
+		const Array &keys = p_attributes.keys();
+		for (int i = 0; i < keys.size(); i++) {
+			const String &key = keys[i];
+			attributes[key] = _as_attribute(p_attributes[key]);
+		}
+	}
+
+	android_plugin->call(ANDROID_SN(log), p_level, body, attributes);
 }
 
 String AndroidSDK::capture_message(const String &p_message, Level p_level) {
@@ -175,7 +234,9 @@ void AndroidSDK::init(const PackedStringArray &p_global_attachments, const Calla
 			SentryOptions::get_singleton()->get_dist(),
 			SentryOptions::get_singleton()->get_environment(),
 			SentryOptions::get_singleton()->get_sample_rate(),
-			SentryOptions::get_singleton()->get_max_breadcrumbs());
+			SentryOptions::get_singleton()->get_max_breadcrumbs(),
+			SentryOptions::get_singleton()->get_experimental()->get_enable_logs(),
+			SentryOptions::get_singleton()->get_experimental()->before_send_log.is_valid() ? before_send_log_handler->get_instance_id() : 0);
 
 	if (is_enabled()) {
 		set_user(SentryUser::create_default());
@@ -202,6 +263,9 @@ AndroidSDK::AndroidSDK() {
 
 	before_send_handler = memnew(SentryAndroidBeforeSendHandler);
 	before_send_handler->_initialize(android_plugin);
+
+	before_send_log_handler = memnew(SentryAndroidBeforeSendLogHandler);
+	before_send_log_handler->_initialize(android_plugin);
 }
 
 AndroidSDK::~AndroidSDK() {
@@ -210,8 +274,11 @@ AndroidSDK::~AndroidSDK() {
 	}
 
 	AndroidStringNames::destroy_singleton();
-	if (before_send_handler != nullptr) {
+	if (before_send_handler) {
 		memdelete(before_send_handler);
+	}
+	if (before_send_log_handler) {
+		memdelete(before_send_log_handler);
 	}
 }