+++ src/sentry/logging/sentry_godot_logger.cpp
@@ -306,6 +306,7 @@ void SentryGodotLogger::_log_error(const String &p_function, const String &p_fil
 	}
 
 	String error_message = p_rationale.is_empty() ? p_code : p_rationale;
+	String error_type = error_type_as_string[int(p_error_type)];
 
 	ErrorKey error_key;
 	error_key.message = error_message;
@@ -316,6 +317,7 @@ void SentryGodotLogger::_log_error(const String &p_function, const String &p_fil
 
 	bool as_event = false;
 	bool as_breadcrumb = false;
+	bool as_log = false;
 
 	{
 		std::lock_guard lock{ error_mutex };
@@ -335,6 +337,8 @@ void SentryGodotLogger::_log_error(const String &p_function, const String &p_fil
 				!is_spammy_error;
 		as_breadcrumb = SentryOptions::get_singleton()->should_capture_breadcrumb((GodotErrorType)p_error_type) &&
 				!is_spammy_error;
+		as_log = SentryOptions::get_singleton()->get_experimental()->get_enable_logs() &&
+				!is_spammy_error;
 
 		if (as_event) {
 			// We decided to capture the error as event (it's happening).
@@ -348,7 +352,7 @@ void SentryGodotLogger::_log_error(const String &p_function, const String &p_fil
 		}
 	}
 
-	if (!as_breadcrumb && !as_event) {
+	if (!as_breadcrumb && !as_event && !as_log) {
 		sentry::logging::print_debug("error capture skipped due to limits");
 		return;
 	}
@@ -356,7 +360,9 @@ void SentryGodotLogger::_log_error(const String &p_function, const String &p_fil
 	sentry::logging::print_debug(
 			"Capturing error: ", error_message,
 			"\n   at: ", p_function, " (", p_file, ":", p_line, ")",
-			"\n   event: ", as_event, "  breadcrumb: ", as_breadcrumb);
+			"\n   event: ", as_event, "  breadcrumb: ", as_breadcrumb, "  log: ", as_log);
+
+	String event_uuid;
 
 	// Capture error as event.
 	if (as_event) {
@@ -375,13 +381,13 @@ void SentryGodotLogger::_log_error(const String &p_function, const String &p_fil
 		Ref<SentryEvent> ev = SentrySDK::get_singleton()->create_event();
 		ev->set_level(sentry::get_sentry_level_for_godot_error_type((GodotErrorType)p_error_type));
 		SentryEvent::Exception exception = {
-			error_type_as_string[int(p_error_type)],
+			error_type,
 			error_message,
 			frames
 		};
 		ev->add_exception(exception);
 		ev->set_logger(logger_name);
-		SentrySDK::get_singleton()->capture_event(ev);
+		event_uuid = SentrySDK::get_singleton()->capture_event(ev);
 	}
 
 	// Capture error as breadcrumb.
@@ -392,7 +398,7 @@ void SentryGodotLogger::_log_error(const String &p_function, const String &p_fil
 		data["line"] = p_line;
 		data["code"] = p_code;
 		data["rationale"] = p_rationale;
-		data["error_type"] = String(error_type_as_string[int(p_error_type)]);
+		data["error_type"] = error_type;
 
 		Ref<SentryBreadcrumb> crumb = SentryBreadcrumb::create(error_message);
 		crumb->set_level(sentry::get_sentry_level_for_godot_error_type((GodotErrorType)p_error_type));
@@ -401,14 +407,48 @@ void SentryGodotLogger::_log_error(const String &p_function, const String &p_fil
 		crumb->set_data(data);
 		SentrySDK::get_singleton()->add_breadcrumb(crumb);
 	}
+
+	// Capture as structured log.
+	if (as_log) {
+		String body = vformat("%s: %s\n   at: %s (%s:%d)",
+				error_type,
+				error_message,
+				p_function,
+				p_file,
+				p_line);
+		if (as_event) {
+			// TODO: Should just leave it as attribute?
+			body += "\n   event_id: " + event_uuid;
+		}
+
+		LogLevel log_level = sentry::get_sentry_log_level_for_godot_error_type((GodotErrorType)p_error_type);
+
+		Dictionary attributes;
+		attributes["error.function"] = p_function;
+		attributes["error.file"] = p_file;
+		attributes["error.line"] = p_line;
+		attributes["error.type"] = error_type;
+		if (!event_uuid.is_empty()) {
+			attributes["sentry.event_id"] = event_uuid;
+		}
+		if (!p_code.is_empty()) {
+			attributes["error.code"] = p_code;
+		}
+		if (!p_rationale.is_empty()) {
+			attributes["error.rationale"] = p_rationale;
+		}
+
+		SentrySDK::get_singleton()->get_internal_sdk()->log(log_level, body, attributes);
+	}
 }
 
 void SentryGodotLogger::_log_message(const String &p_message, bool p_error) {
 	sentry::logging::MessageScope message_scope;
 
+	bool as_log = SentryOptions::get_singleton()->get_experimental()->get_enable_logs();
 	bool as_breadcrumb = SentryOptions::get_singleton()->is_logger_messages_as_breadcrumbs_enabled();
 
-	if (!as_breadcrumb) {
+	if (!as_log && !as_breadcrumb) {
 		return;
 	}
 
@@ -432,6 +472,11 @@ void SentryGodotLogger::_log_message(const String &p_message, bool p_error) {
 		if (processed_message.begins_with(prefix)) {
 			return;
 		}
+	}
+
+	if (as_log) {
+		sentry::LogLevel level = p_error ? LOG_LEVEL_ERROR : LOG_LEVEL_INFO;
+		SentrySDK::get_singleton()->get_internal_sdk()->log(level, processed_message, log_attributes);
 	}
 
 	if (as_breadcrumb) {
@@ -467,6 +512,8 @@ void SentryGodotLogger::_notification(int p_what) {
 
 SentryGodotLogger::SentryGodotLogger() {
 	logger_name = "SentryGodotLogger";
+
+	log_attributes["sentry.origin"] = "auto.log.godot";
 
 	// Filtering setup.
 	filter_by_prefix = {